api/requirements.txt
azure-functions
black